pyproject.toml
[project]
name = "toy-translator"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.2.1",
]
